Move NulTerminatedBytes to its own crate
