+++ src/redisearch_rs/Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "index_result",
     "inverted_index",
     "inverted_index_bencher",
+    "nul_terminated_bytes",
     "numeric_range_tree",
     "thin_vec",
     "qint",
@@ -129,6 +130,7 @@ index_spec = { path = "./c_wrappers/index_spec" }
 index_result = { path = "./index_result" }
 index_spec_cache = { path = "./c_wrappers/index_spec_cache" }
 inverted_index = { path = "./inverted_index" }
+nul_terminated_bytes = { path = "./nul_terminated_bytes" }
 numeric_range_tree = { path = "./numeric_range_tree" }
 numeric_score_source = { path = "./numeric_score_source" }
 redis_reply = { path = "./redis_reply" }
