+++ src/redisearch_rs/c_entrypoint/fork_gc_ffi/Cargo.toml
@@ -19,7 +19,7 @@ cheadergen.workspace = true
 ffi = { workspace = true }
 fork_gc = { workspace = true }
 index_spec.workspace = true
-string_utils.workspace = true
+nul_terminated_bytes.workspace = true
 tracing.workspace = true
 tracing_log_error.workspace = true
 workspace_hack.workspace = true
