+++ src/redisearch_rs/fork_gc/Cargo.toml
@@ -30,7 +30,7 @@ redis-module.workspace = true
 rmp-serde.workspace = true
 index_spec.workspace = true
 serde.workspace = true
-string_utils.workspace = true
+nul_terminated_bytes.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 workspace_hack.workspace = true
