+++ src/redisearch_rs/nul_terminated_bytes/Cargo.toml
@@ -1,13 +1,11 @@
 [package]
-name = "query_term"
+name = "nul_terminated_bytes"
 version.workspace = true
 edition.workspace = true
 license-file.workspace = true
 publish.workspace = true
 
 [dependencies]
-cheadergen.workspace = true
-nul_terminated_bytes.workspace = true
 workspace_hack.workspace = true
 
 [lints]
