+++ src/redisearch_rs/string_utils/src/lib.rs
@@ -12,9 +12,6 @@
 //! These are pure-Rust replacements for C helpers that were previously
 //! implemented using `libnu` for Unicode operations.
 
-pub mod nul_terminated_bytes;
 pub mod runes;
 pub mod tag;
 pub mod unicode;
-
-pub use nul_terminated_bytes::NulTerminatedBytes;
