+++ src/redisearch_rs/string_utils/tests/integration/main.rs
@@ -11,7 +11,6 @@
 redis_mock::mock_or_stub_missing_redis_c_symbols!();
 extern crate redisearch_rs;
 
-mod nul_terminated_bytes;
 mod str_to_lower_runes;
 mod tag_strtolower;
 mod unicode_tolower;
