+++ src/redisearch_rs/value/Cargo.toml
@@ -11,7 +11,7 @@ cheadergen.workspace = true
 ffi.workspace = true
 libc.workspace = true
 query_error.workspace = true
-string_utils.workspace = true
+nul_terminated_bytes.workspace = true
 workspace_hack.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
